ehazira-backend/requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# Authentication
PyJWT==2.8.0
google-auth==2.25.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Redis (for channels and cache)
redis==5.0.1

# CORS headers
django-cors-headers==4.3.1

# Push notifications
pyfcm==1.5.4  # Firebase Cloud Messaging

# ASGI server
daphne==4.0.0

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Excel file support
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1

# Code quality
black==23.12.1
flake8==6.1.0

firebase-admin==6.3.0